First steps in refactoring. Glance and Keystone
Added SSL to Keystone. Merged all Glance containers into one
service container. The added complexity makes for what I hope to
be a simpler user experience. The service containers hold all of
the configuration and software needed to initialize and run the
servers, rather than using run-once configuration containers. Had
to build some dubious logic to make the initialization idempotent,
but appears to work in general. Will apply this to the other
services in an effort to simplify the entire build-configure-startup
process.